Support multiple parallel writers
This requires thread-save HDF5 to be available, which is a bit of a mixed
bag. There still seems to be a residual chance of it failing to create
the initial files / data sets, not quite sure why.

Behind the scenes this needs quite a bit of refactoring. Especially note
that the writer threads are now started using pthreads again, because we
cannot use OpenMP sections to start a variable number of threads (sigh).

This also introduces tagged statsd statistics (following the DogStatsD
extension) so statistics from different writer threads can be
distinguished.