Some refactoring of chunk/subgrid overlap checks
The idea is that instead of making the decision about whether to
negate a coordinate per visibility, we make it per chunk. This means
that any given chunk can only overlap *either* subgrid 0/1 or 0/-1,
but not both. This substantially reduces the number of rewritten
chunks - only downside is that for hypothetical massive chunk sizes we
might now actually get overlaps for subgrids with negative u.