Refactor to allow better gridder implementation
The first big changes here are that we separate the (sub-)grid size from
the stride, as it can cause L1 cache thrashing when the gridding
kernel attempts to move by steps that are a multiple of the L1 cache
size. This involves a bit of rearranging.

Secondly we shift the gridding interface a bit so that we are always
degridding an entire line of equidistant visibilities. This allows us
to do a bunch of low-level optimisations, most notably it cuts down on
the number of bound checks we can do, as we can easily derive which
visibilities are in-bounds and which are not.