Use small sleeps to reduce CPU usage
There's two places where we would like to yield the CPU, but clearly
aren't very successful at it: When we block on the subgrid queue, and
we are waiting for other MPI ranks to finish. In either case we don't
care too much about getting woken up immediately, so let's just use
usleep until we have a better solution available.