AVX2 kernel generation
Using a small code generator in Python seems like a decent way to
specialise kernels. This specifically creates kernels that retain both
parts of the convolution kernel in registers, with the inner direction
in "duplicated" layout so that doing the multiplication is a single
instruction.

This also pre-fetches the kernel for the *next* visibility while the
current one is getting worked on, in an attempt to reduce the effect
of latencies. Doesn't make for the most readable code, but is
approaching about 50% efficiency relative to linpack, which is pretty
good given that for the inner loop we are accessing unaligned memory
basically every single instruction.