Working w-wstacks
The overall least efficient way of dealing with non-coplanarity we are
going to support: Multiplying in the Fresnel pattern on the facet (the
equivalent of classic w-stacking). Also as usual this only allows us
to get close to the w-levels we want, so the visibility accuracy
checks will not succeed just yet.

On the positive side - this is relatively simple to implement, so
gives us a chance to get all the required refactoring out of the
way. Especially note that there is now a sequential outer loop for
walking through w-levels. Not entirely sure it's optimal to go from
minimal w-value to maximum, might have to be revisited at some point.