Refactor to use bl_data in configuration
There seems to be little reason not to calculate all UVWs up-front,
this means that we can just pass cached bl_data around and have it
contain all the information we might want to have on a baseline. This
especially means we *finally* get to use bl_bounding_box in the
streamer.