Move subgrid FFT to workers
This is redundant work especially given that for the quality checks we
want the data in grid space... Yet it's the "only way" given that for
w-towers we will want to multiply in the Fresnel pattern in image
space, so having the data start out in grid space is awkward.

Maybe there are some shrewd caching/work sharing schemes here, yet all
of that will go out of the window anyway once at some (far off?) point
we get baseline- and direction-dependent effects to correct
for. There's going to be so much FFT work at that point that we better
get into the habit of not stressing too much about it.